+++ src/hooks/index.ts
@@ -12,3 +12,4 @@ export {useRenderForcingRef} from './useRenderForcingRef'
 export {useProvidedStateOrCreate} from './useProvidedStateOrCreate'
 export {useMenuInitialFocus} from './useMenuInitialFocus'
 export {useTypeaheadFocus} from './useTypeaheadFocus'
+export {useMenuKeyboardNavigation} from './useMenuKeyboardNavigation'
